d2sstats: use 64-bit values/support 64-bit custom user data; readme: fix medianXL configuration

+++ pkg/d2s/api.go
@@ -111,7 +111,7 @@ func (d *D2S) SetLevel(level byte) *D2S {
 	}
 
 	d.Level = level
-	d.Stats.Level = uint32(level)
+	d.Stats.Level = uint64(level)
 
 	return d
 }
